Move Timeline files into their own package
