+++ app/src/androidTest/java/com/keylesspalace/tusky/TimelineDAOTest.kt
@@ -5,7 +5,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import com.keylesspalace.tusky.db.*
 import com.keylesspalace.tusky.entity.Status
-import com.keylesspalace.tusky.repository.TimelineRepository
+import com.keylesspalace.tusky.components.timeline.TimelineRepository
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
