+++ app/src/main/java/com/keylesspalace/tusky/TabData.kt
@@ -21,7 +21,7 @@ import androidx.annotation.StringRes
 import androidx.fragment.app.Fragment
 import com.keylesspalace.tusky.components.conversation.ConversationsFragment
 import com.keylesspalace.tusky.fragment.NotificationsFragment
-import com.keylesspalace.tusky.fragment.TimelineFragment
+import com.keylesspalace.tusky.components.timeline.TimelineFragment
 
 /** this would be a good case for a sealed class, but that does not work nice with Room */
